Feature:  New
Add the ability to save a model to disk as an HDF5 file.  This means
that testing can be done by a separate program that loads the
model and the new data.  While this was done to address memory constraints
as GPU resources were not being released after training, so that
testing led to out-of-memory (OOM) errors, it is beneficial in that
the same model can then be loaded and run with different data sets
without having to run through the training every time.